+++ src/app/services/file-sharing/core-file-sharing.service.ts
@@ -1,0 +1,9 @@
+import { Injectable } from '@angular/core';
+
+@Injectable({
+  providedIn: 'root'
+})
+export class CoreFileSharingService {
+
+  constructor() { }
+}
